requirements.txt
appdirs==1.4.3
asgiref==3.2.5
attrs==19.3.0
black==19.10b0
click==7.1.1
Django==3.0.4
django-countries==6.0
pathspec==0.7.0
pytz==2019.3
regex==2020.2.20
rstr==2.2.6
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
django-tastypie
